Revert "Merge remote-tracking branch 'upstream/master' into SHARPFONT-VER-2-12-1"
This reverts commit 03f6caa71120103afe2c46d217d1d2b1d78d02b5.

+++ src/truetype/ttobjs.c
@@ -1004,7 +1004,7 @@
     {
       size->cvt[i] = FT_MulFix( face->cvt[i], scale );
       FT_TRACE6(( "  %3d: %f (%f)\n",
-                  i, (double)face->cvt[i] / 64, (double)size->cvt[i] / 64 ));
+                  i, face->cvt[i] / 64.0, size->cvt[i] / 64.0 ));
     }
     FT_TRACE6(( "\n" ));
